requirements.txt
numpy==1.24.3
matplotlib==3.7.1
python-docx==0.8.11
